% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-definitions.R
\docType{data}
\name{demoFreq}
\alias{demoFreq}
\title{Demo dataset with Words and Frequency}
\format{A data set with 1011 observations of 2 variables, words and frequancy}
\usage{
demoFreq
}
\description{
A data file of words and frequency from tm package
}
\keyword{datasets}

