% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-definitions.R
\docType{data}
\name{demoFreqC}
\alias{demoFreqC}
\title{Demo dataset with Chinese character Words and Frequency}
\format{
A data set with 885 observations of 2 variables, words and frequancy
}
\usage{
demoFreqC
}
\description{
A data file of words and frequency from tm package
}
\keyword{datasets}
