% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/letterCloud.R
\name{letterCloud}
\alias{letterCloud}
\title{Create wordcloud with the shape of a word}
\usage{
letterCloud(data, word, wordSize = 0, letterFont = NULL, ...)
}
\arguments{
\item{data}{A data frame including word and freq in each column}

\item{word}{A word to create shape for wordcloud.}

\item{wordSize}{Parameter of the size of the word.}

\item{letterFont}{Letter font}

\item{...}{Other parameters for wordcloud.}
}
\description{
Function for Creating wordcloud with the shape of a word
}
\examples{
library(wordcloud2)

letterCloud(demoFreq,"R")
}

