% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plot-construction.R
\name{+.wordcloud2}
\alias{+.wordcloud2}
\title{Change the themes of wordcloud2}
\usage{
\method{+}{wordcloud2}(e1, e2)
}
\arguments{
\item{e1}{wordcloud2 object}

\item{e2}{wordcloud2 themes}
}
\description{
Function for Creating wordcloud theme
}
\examples{
wc = wordcloud2(demoFreq)

wc + WCtheme(1)
wc + WCtheme(2)
wc + WCtheme(3)
wc + WCtheme(2) + WCtheme(3)
}

