% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme.R
\name{WCtheme}
\alias{WCtheme}
\title{Change the themes of wordcloud2}
\usage{
WCtheme(class = 1)
}
\arguments{
\item{class}{class for theme in wordcloud2}
}
\description{
Function for Creating wordcloud theme
}
\examples{
wc = wordcloud2(demoFreq)

wc + WCtheme(1)
wc + WCtheme(2)
wc + WCtheme(3)
wc + WCtheme(2) + WCtheme(3)
}
