% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/renderWordcloud2.R
\name{wordcloud2-shiny}
\alias{wordcloud2-shiny}
\alias{wordcloud2Output}
\alias{renderWordcloud2}
\title{Plot wordcloud2 in shiny}
\usage{
wordcloud2Output(outputId, width = "100\%", height = "400px",
                 clickedWordInputId = paste0(outputId, "_clicked"))
renderWordcloud2(expr, env = parent.frame(), quoted = FALSE)

wordcloud2Output(outputId, width = "100\%", height = "400px",
  clickedWordInputId = paste0(outputId, "_clicked"))

renderWordcloud2(expr, env = parent.frame(), quoted = FALSE)
}
\arguments{
\item{outputId}{output variable to read from}

\item{width, height}{Must be a valid CSS unit (like \code{"100\%"},
\code{"400px"}, \code{"auto"}) or a number, which will be coerced to a
string and have \code{"px"} appended.}

\item{clickedWordInputId}{The input id to assign the value (see details) of the clicked word.
Defaults to \code{outputId_clicked} and can be referenced in the server as \code{input$outputId_clicked}.}

\item{expr}{An expression that generates a networkD3 graph}

\item{env}{The environment in which to evaluate \code{expr}.}

\item{quoted}{Is \code{expr} a quoted expression (with \code{quote()})? This
is useful if you want to save an expression in a variable.}
}
\description{
Function for plotting wordcloud2 in shiny
}
\details{
Use renderWordcloud2 to render an wordcloud2 object and use wordcloud2Output
output an wordcloud2 object. See more details in shiny package.

The value assigned to the associated \code{clickedWordInputId} will be of the form \code{word:freq}.
This will be the same value as shown in the hover information (example: \code{super:32}).
}
