% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/renderWordcloud2.R
\name{wordcloud2Output}
\alias{wordcloud2Output}
\title{Plot wordcloud2 in shiny}
\usage{
wordcloud2Output(outputId, width = "100\%", height = "400px")
renderWordcloud2(expr, env = parent.frame(), quoted = FALSE)
}
\description{
Function for plotting wordcloud2 in shiny
}
\details{
Use renderWordcloud2 to render an wordcloud2 object and use wordcloud2Output
output an wordcloud2 object. See more details in shiny package.
}
\examples{
library(shiny)
library(wordcloud2)
# Global variables can go here
n <- 1

# Define the UI
ui <- bootstrapPage(
 numericInput('size', 'Size of wordcloud', n),
 wordcloud2Output('wordcloud2')
)


# Define the server code
server <- function(input, output) {
 output$wordcloud2 <- renderWordcloud2({
   # wordcloud2(demoFreqC, size=input$size)
   wordcloud2(demoFreq, size=input$size)
 })
}
# Return a Shiny app object
# Sys.setlocale("LC_CTYPE","chs") #if you use Chinese
shinyApp(ui = ui, server = server)
}

